'C-x C-k' as conventional key to kill completion items
Added this binding to read_buffer_keymap for read-buffer-kill-buffer,
while also leaving 'C-k' bound to that command, since the conflict with
the 'C-k' of text editing is not important in this context.

Bound 'C-x C-k' to read-url-kill-item in read_url_keymap, and removed
'C-k' binding there, because people expect plain old C-k to kill-line in
that context.